+++ lib/flutter_audio_waveforms.dart
@@ -2,6 +2,7 @@ export 'package:flutter_audio_waveforms/src/core/audio_waveform.dart';
 export 'package:flutter_audio_waveforms/src/core/waveform_painters_ab.dart';
 export 'package:flutter_audio_waveforms/src/util/waveform_alignment.dart';
 export 'package:flutter_audio_waveforms/src/waveforms/curved_polygon_waveform/curved_polygon_waveform.dart';
+export 'package:flutter_audio_waveforms/src/waveforms/frequency_graph.dart';
 export 'package:flutter_audio_waveforms/src/waveforms/polygon_waveform/polygon_waveform.dart';
 export 'package:flutter_audio_waveforms/src/waveforms/rectangle_waveform/rectangle_waveform.dart';
 export 'package:flutter_audio_waveforms/src/waveforms/squiggly_waveform/squiggly_waveform.dart';