+++ lib/src/waveforms/rectangle_waveform/inactive_waveform_painter.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: public_member_api_docs
+
 import 'package:flutter/material.dart';
 import 'package:flutter_audio_waveforms/src/core/waveform_painters_ab.dart';
 import 'package:flutter_audio_waveforms/src/util/waveform_alignment.dart';
@@ -14,6 +16,8 @@ class RectangleInActiveWaveformPainter extends InActiveWaveformPainter {
     required double sampleWidth,
     required Color borderColor,
     required double borderWidth,
+    required this.isRoundedRectangle,
+    required this.isCentered,
   }) : super(
           samples: samples,
           color: color,
@@ -24,6 +28,9 @@ class RectangleInActiveWaveformPainter extends InActiveWaveformPainter {
           borderWidth: borderWidth,
           style: PaintingStyle.fill,
         );
+
+  final bool isRoundedRectangle;
+  final bool isCentered;
 
   @override
   void paint(Canvas canvas, Size size) {
@@ -39,21 +46,108 @@ class RectangleInActiveWaveformPainter extends InActiveWaveformPainter {
       ..strokeWidth = borderWidth;
     //Gets the [alignPosition] depending on [waveformAlignment]
     final alignPosition = waveformAlignment.getAlignPosition(size.height);
+    final isAbsolute = waveformAlignment != WaveformAlignment.center;
 
+    if (isRoundedRectangle) {
+      drawRoundedRectangles(
+        canvas,
+        alignPosition,
+        paint,
+        borderPaint,
+        waveformAlignment,
+        isCentered,
+        isAbsolute,
+      );
+    } else {
+      drawRegularRectangles(
+        canvas,
+        alignPosition,
+        paint,
+        borderPaint,
+        waveformAlignment,
+        isCentered,
+        isAbsolute,
+      );
+    }
+  }
+
+  // ignore: long-parameter-list
+  void drawRegularRectangles(
+    Canvas canvas,
+    double alignPosition,
+    Paint paint,
+    Paint borderPaint,
+    WaveformAlignment waveformAlignment,
+    bool isCentered,
+    bool isAbsolute,
+  ) {
+    final isCenteredAndNotAbsolute = isCentered && !isAbsolute;
     for (var i = 0; i < samples.length; i++) {
       final x = sampleWidth * i;
-      final y = samples[i];
+      final y = isCenteredAndNotAbsolute ? samples[i] * 2 : samples[i];
+      final positionFromTop =
+          isCenteredAndNotAbsolute ? alignPosition - y / 2 : alignPosition;
+      final rectangle = Rect.fromLTWH(x, positionFromTop, sampleWidth, y);
+
       //Draws the filled rectangles of the waveform.
       canvas
         ..drawRect(
-          Rect.fromLTWH(x, alignPosition, sampleWidth, y),
+          rectangle,
           paint,
         )
         //Draws the border for the rectangles of the waveform.
         ..drawRect(
-          Rect.fromLTWH(x, alignPosition, sampleWidth, y),
+          rectangle,
           borderPaint,
         );
     }
+  }
+
+  // ignore: long-parameter-list
+  void drawRoundedRectangles(
+    Canvas canvas,
+    double alignPosition,
+    Paint paint,
+    Paint borderPaint,
+    WaveformAlignment waveformAlignment,
+    bool isCentered,
+    bool isAbsolute,
+  ) {
+    final radius = Radius.circular(sampleWidth);
+    final isAbsoluteAndNotCentered = isAbsolute || !isCentered;
+    for (var i = 0; i < samples.length; i++) {
+      if (i.isEven) {
+        final x = sampleWidth * i;
+        final y = isAbsoluteAndNotCentered ? samples[i] : samples[i] * 2;
+        final positionFromTop =
+            isAbsoluteAndNotCentered ? alignPosition : alignPosition - y / 2;
+        final rectangle = Rect.fromLTWH(x, positionFromTop, sampleWidth, y);
+        //Draws the filled rectangles of the waveform.
+        canvas
+          ..drawRRect(
+            RRect.fromRectAndRadius(
+              rectangle,
+              radius,
+            ),
+            paint,
+          )
+          //Draws the border for the rectangles of the waveform.
+          ..drawRRect(
+            RRect.fromRectAndRadius(
+              rectangle,
+              radius,
+            ),
+            borderPaint,
+          );
+      }
+    }
+  }
+
+  @override
+  bool shouldRepaint(covariant RectangleInActiveWaveformPainter oldDelegate) {
+    // TODO: implement shouldRepaint
+    return getShouldRepaintValue(oldDelegate) ||
+        isRoundedRectangle != oldDelegate.isRoundedRectangle ||
+        isCentered != oldDelegate.isCentered;
   }
 }